♻️ Move id_gen_chacha8.go to opts
