+++ id_gen.go
@@ -5,14 +5,11 @@ import "github.com/akm/sql-slog/opts"
 // IDGen is a function that generates an ID string.
 type IDGen = opts.IDGen
 
+var IDGeneratorDefault = opts.IDGeneratorDefault
+
 // IDGenerator returns an Option that sets the ID generator.
 // The default is IDGeneratorDefault.
 func IDGenerator(idGen IDGen) Option { return func(o *Options) { o.IDGen = idGen } }
-
-const defaultIDLength = 16
-
-// IDGeneratorDefault is the default ID generator.
-var IDGeneratorDefault = NewChaCha8IDGenerator(defaultIDLength).Generate
 
 // ConnIDKey sets the key for the connection ID.
 // The default is ConnIDKeyDefault.
