+++ opts/id_gen_chacha8.go
@@ -1,4 +1,4 @@
-package sqlslog
+package opts
 
 import (
 	cryptoRand "crypto/rand"
